Created models, routes, controllers for backend
I was most accustomed to using these technologies for the backend. I was not too sure if we have or had any constraints with the amount of dependencies we want to use.

The only one that I've implemented was fortuneMsg, which holds the fortune message, and it is successfully displayed in saved readings.