fix: set local endpoint for server to receive post requests
In this commit I'm connecting to my local database in `server.js`. Also, the current post endpoint for forms are under localhost. We need to change this after we deploy the server.

BREAKING CHANGE: if you have a database created with the old schema, you need to drop the `users` collection, or else there will be bugs involving the uniqueness of fields that doesn't even exist anymore.